nmap-requirements.txt
# Core dependencies
python-nmap>=0.7.1
argparse>=1.4.0

# Enhanced features
tqdm>=4.65.0
ipaddress>=1.0.23
asyncio>=3.4.3
typing-extensions>=4.5.0

# Output formatting
json5>=0.9.14
csv23>=0.3.4

# Logging and error handling
logging>=0.4.9.6

# Testing and development
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
flake8>=6.0.0
black>=23.3.0

# Documentation
sphinx>=7.0.1
sphinx-rtd-theme>=1.2.0
